# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DeguBase – workspace and table management API.",
        "title": "DeguBase API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/workspaces/{wsCode}/skills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "List skills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Skill"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Create skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "skill",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skills.CreateSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/skills/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/markdown"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Preview skill (without persisting)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "skill preview params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skills.PreviewSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/skills/{skillID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Get skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skill ID",
                        "name": "skillID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Delete skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skill ID",
                        "name": "skillID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Update skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skill ID",
                        "name": "skillID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "skill",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skills.UpdateSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/skills/{skillID}/context": {
            "get": {
                "produces": [
                    "text/markdown"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Skill context (LLM-ready)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skill ID",
                        "name": "skillID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{wsCode}/skills/{skillID}/skill.md": {
            "get": {
                "produces": [
                    "text/markdown"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Download skill as Markdown",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "skill ID",
                        "name": "skillID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "models.Skill": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token_id": {
                    "type": "integer"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
...
        "skills.CreateSkillRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Access workspace data"
                },
                "name": {
                    "type": "string",
                    "example": "My Skill"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "list",
                        "create",
                        "update",
                        "delete"
                    ]
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token_name": {
                    "type": "string",
                    "example": "my-skill-token"
                }
            }
        },
        "skills.PreviewSkillRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "skills.UpdateSkillRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Access workspace data"
                },
                "name": {
                    "type": "string",
                    "example": "My Skill"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "list",
                        "create",
                        "update",
                        "delete"
                    ]
                },
                "table_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}